Re-organize namespaces
