+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Application/ChangeOfSupplier/Processing/ConsumerDetails/ForwardConsumerDetails.cs
@@ -12,24 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.MarketRoles.Domain.MeteringPoints;
-using Energinet.DataHub.MarketRoles.Domain.SeedWork;
+using System;
+using Energinet.DataHub.MarketRoles.Application.Common.Commands;
 
-namespace Energinet.DataHub.MarketRoles.Application.ChangeOfSupplier.Processing.Events
+namespace Energinet.DataHub.MarketRoles.Application.ChangeOfSupplier.Processing.ConsumerDetails
 {
-    public class ConsumerDetailsDispatched : DomainEventBase
+    public class ForwardConsumerDetails : InternalCommand
     {
-        public ConsumerDetailsDispatched(AccountingPointId accountingPointId, BusinessProcessId businessProcessId, Transaction transaction)
+        public ForwardConsumerDetails(Guid accountingPointId, Guid businessProcessId, string transaction)
         {
             AccountingPointId = accountingPointId;
             BusinessProcessId = businessProcessId;
             Transaction = transaction;
         }
 
-        public AccountingPointId AccountingPointId { get; }
+        public Guid AccountingPointId { get; }
 
-        public BusinessProcessId BusinessProcessId { get; }
+        public Guid BusinessProcessId { get; }
 
-        public Transaction Transaction { get; }
+        public string Transaction { get; }
     }
 }
