+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Infrastructure/EDIMessaging/ENTSOE/CIM/ChangeOfSupplier/MeteringPointDetails/MeteringPointDetailsForwarder.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using System.Threading.Tasks;
-using Energinet.DataHub.MarketRoles.Application.ChangeOfSupplier.Processing.Commands.MeteringPointDetails;
+using Energinet.DataHub.MarketRoles.Application.ChangeOfSupplier.Processing.MeteringPointDetails;
 using Energinet.DataHub.MarketRoles.Domain.MeteringPoints;
 
 namespace Energinet.DataHub.MarketRoles.Infrastructure.EDIMessaging.ENTSOE.CIM.ChangeOfSupplier.MeteringPointDetails
